LayoutTests/imported/w3c/web-platform-tests/custom-elements/scoped-registry/constructor-reentry-with-different-definition.tentative-expected.txt

CONSOLE MESSAGE: TypeError: Cannot instantiate a custom element inside its own constructor during upgrades
CONSOLE MESSAGE: TypeError: Cannot instantiate a custom element inside its own constructor during upgrades

PASS Re-entry via upgrade before calling super()
PASS Re-entry via upgrade after calling super()
PASS Re-entry via direct constructor call before calling super()
PASS Re-entry via direct constructor call after calling super()
